requirements.txt
streamlit==1.34.0
pytesseract==0.3.10
streamlit-cropper==0.2.2
pillow==10.3.0
st-pages==0.4.5
st-social-media-links==0.1.1